Add Object class to example
Under cmake, make the default glue targets invalid. Assert that the glue
targets are valid iff they're required. Cmake will now respect the
OMR_EXAMPLE flag, which will force set the glue to the example code.
This flag is on by default. The end result is, if you disable the
example, and fail to provide your own glue, the build will complain.
This sets up OMR for multiple consumers.

Begin building the example under cmake. The example currently hangs, so
we can't run it as a test. This is an old issue, unrelated to this PR.
Fix up linking with the glue.

Add a new include directory to the example. This allows us to scope our
example headers under the OMR/App directory. Eventually, all example
headers and glue code will move under this directory.

Begin using the OMR::App namespace for new application-specific code.
Eventually, all glue will be defined under this name.

Add a new class, Object, which represents the data stored in an object.
Update the omrobjectptr_t class to point to the object. Update the
object model to use this type.

Signed-off-by: Robert Young <rwy0717@gmail.com>