Add IntrusiveList utility
The IntrusiveList is a doubly linked list, where the node is stored in
the elements itself. The node access mechanism is overrideable, allowing
elements to customize node access, or belong in multiple lists
simultaneously.

Compared to the J9_LINKED_LIST, this utility is implemented without
macros while still being type-safe.

Signed-off-by: Robert Young <rwy0717@gmail.com>