Add basic type_traits-like utility to OMR
Type traits are useful in templated code to make queries or assertions
about type parameters. For example, a container might assert that it
never contains a reference type.  A smart pointer won't define
pointer-operators when it's referent is void.

Our supported compilers don't provide the standard C++11 header
<type_traits> in OMR. Even looking at planned compiler upgrades, we'll
still be missing type_traits on XLC, and it's unlikely to ever be
provided.

This header provides some easily implemented type traits without any
hacks or compiler support. As well, this header follows OMR coding
conventions, rather than try to simulate the standard type_traits
header.

This PR implements the following:

* Static constants:
  * IntegralConstant<T, X>
  * BoolConstant<B> / TrueConstant / FalseConstant
  * TypeAlias<T>
* types queries:
  * IsSame<T, U>
  * IsReference<T>
  * IsVoid<T>
  * IsPointer<T>
* type modifications:
  * RemoveConst<T>
  * RemoveVolatile<T>
  * RemoveCv<T>
  * RemoveReference<T>
  * RemoveCvRef<T>

There's lots of opportunities for expanding the traits provided in this
initial commit, but this is a good base for defining container types in
OMR.

Signed-off-by: Robert Young <rwy0717@gmail.com>