Add OMR::GC::Handle<T> value type
Handle<T> is a pointer-like value that indirects through an externally
rooted slot. This allows users to pass rooted gc-references through
function calls, independently of how the object is actually rooted.  The
main utility of this class is that it _signifies_ that the caller has
rooted the parameters of a function.

It is possible for Handle<T>'s to outlive their roots, which is an
extremely dangerous error. In the future, we may chose to add "handle
counting" to specialized root structures, to ensure there are no
outstanding handles when a root dies.

Another potential future enhancement would be to parameterize the root
type, to allow for different gc-reference encodings. The value of doing
this is questionable though, since the main function of the handle is to
hide the rooting mechanism, and doing this puts it right into the type.

Signed-off-by: Robert Young <rwy0717@gmail.com>