Add RefSlotHandle types to GC
The RefSlotHandle is "A handle to a reference-slot in an object". The
RefSlotHandle is a default implementation of a SlotHandle for bare,
unencoded gc-references.

This PR also adds a CompressedRefSlotHandle, for working with compressed
references. The reference shift is dynamic.

In the future, these classes will be used by OMR client's object
scanners to give the GC read-write capabilities on objects.

Signed-off-by: Robert Young <rwy0717@gmail.com>