Permanently enable OMR_GC_SCAVENGER_DELEGATE
Effectively turn on the OMR_GC_SCAVENGER_DELEGATE flag. This PR removes
any code ifdef'd under this flag was rewritten to assume the flag is
enabled, and dead code was deleted.

The flag remains in the build system temporarily, as a settable option,
but should have no effect on the source code.

Signed-off-by: Robert Young <rwy0717@gmail.com>