Return a reference to a dummy address variable
This code is unimplemented, and should crash before trying to return a
dummy result. Unfortunately, the dummy result is created by
dereferencing a null pointer, and clang will still emit a warning that
this is undefined behaviour. This patch causes the function to return a
valid reference to a global dummy value, which silences the error.

Signed-off-by: Robert Young <rwy0717@gmail.com>