Include <netinet/in.h> in omrsock.h
This header is responsible for defining the in6_addr struct. This type
is used in the header <netinet/tcp.h>. On some versions of AIX ,tcp.h
does not include in.h (despite using types defined there), so we have to
manually include it before tcp.h. This isn't a problem on other posix
platforms.

Signed-off-by: Robert Young <rwy0717@gmail.com>